# optimizer: allow inlining of constant `::TypeName` object