This allows us to inline `Ref.body.name` and such. This doesn't seem to have much impact on sysimage size:

> master
```
➜ ls -la ./usr/lib/julia
total 906640
drwxr-xr-x    7 aviatesk  staff        224 Apr  2 20:51 .
drwxr-xr-x  276 aviatesk  staff       8832 Apr  1 11:22 ..
-rw-r--r--    1 aviatesk  staff   18344432 Apr  2 20:49 corecompiler.ji
-rw-r--r--    1 aviatesk  staff  192771632 Apr  2 20:51 sys-o.a
-rwxr-xr-x    1 aviatesk  staff  161034648 Apr  2 20:51 sys.dylib
drwxr-xr-x    3 aviatesk  staff         96 Apr  1 11:26 sys.dylib.dSYM
-rw-r--r--    1 aviatesk  staff   92039864 Apr  2 20:50 sys.ji
```

> this commit
```
➜ ls -la ./usr/lib/julia
total 906912
drwxr-xr-x    7 aviatesk  staff        224 Apr  2 20:52 .
drwxr-xr-x  276 aviatesk  staff       8832 Apr  2 20:48 ..
-rw-r--r--    1 aviatesk  staff   18329400 Apr  2 20:49 corecompiler.ji
-rw-r--r--    1 aviatesk  staff  192790232 Apr  2 20:52 sys-o.a
-rwxr-xr-x    1 aviatesk  staff  161051240 Apr  2 20:52 sys.dylib
drwxr-xr-x    3 aviatesk  staff         96 Mar 30 13:18 sys.dylib.dSYM
-rw-r--r--    1 aviatesk  staff   92160128 Apr  2 20:50 sys.ji
```

@nanosoldier `runbenchmarks("inference", vs=":master")`