@aviatesk 2023-04-02
@nanosoldier `runbenchmarks("inference", vs=":master")`