@nanosoldier 2023-04-02
[Your benchmark job](https://github.com/JuliaLang/julia/pull/49225#issuecomment-1493359036) has completed - no performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/84a76e1_vs_a612220/report.md).